marzneshin/pyproject.toml
[project]
name = "marzneshin-migration"
version = "0.1.0"
description = "Database migration tool for migrating data from Marzneshin to Pasarguard"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pymysql>=1.0.0",
    "python-dotenv>=1.0.0",
    "xxhash>=3.0.0",
]

[project.optional-dependencies]
dev = []

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["migration"]

